Move FrameBrowser back to its own module, implement optional labels
- FrameBrowser class back out of ArrayTableWidget
- implement labels in ArrayTableWidget.setArrayData:
    - by default, no labels at all (backward compatible API)
    - if labels=True, put default label: "dimension %d"
    - if labels is a list of strings, use the stnigs
    - when perspective is changed and the browsers apply to different dimensions, update the label accordingly (reorder them)
