+++ silx/gui/widgets/FrameBrowser.py
@@ -1,0 +1,239 @@
+# coding: utf-8
+# /*##########################################################################
+#
+# Copyright (c) 2016 European Synchrotron Radiation Facility
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+#
+# ###########################################################################*/
+"""This module defines two main classes:
+
+    - :class:`FrameBrowser`: a widget with 4 buttons (first, previous, next,
+     last) to browse between frames and a text entry to access aspecific frame
+     by typing it's number)
+    - :class:`HorizontalSliderWithBrowser`: a FrameBrowser with an additional
+     slider.
+
+"""
+from silx.gui import qt
+from silx.gui import icons
+
+__authors__ = ["V.A. Sole", "P. Knobel"]
+__license__ = "MIT"
+__date__ = "21/10/2016"
+
+
+icon_first = icons.getQIcon("first")
+icon_previous = icons.getQIcon("previous")
+icon_next = icons.getQIcon("next")
+icon_last = icons.getQIcon("last")
+
+
+class HorizontalSpacer(qt.QWidget):
+    def __init__(self, parent=None):
+        qt.QWidget.__init__(self, parent)
+        self.setSizePolicy(qt.QSizePolicy(qt.QSizePolicy.Expanding,
+                                          qt.QSizePolicy.Fixed))
+
+
+class FrameBrowser(qt.QWidget):
+    """Frame browser widget, with 4 buttons/icons and a line edit to select
+    a frame number in a stack of images."""
+    sigIndexChanged = qt.pyqtSignal(object)
+
+    def __init__(self, parent=None, n=1):
+        qt.QWidget.__init__(self, parent)
+        self.mainLayout = qt.QHBoxLayout(self)
+        self.mainLayout.setContentsMargins(0, 0, 0, 0)
+        self.mainLayout.setSpacing(0)
+        self.firstButton = qt.QPushButton(self)
+        self.firstButton.setIcon(icon_first)
+        self.previousButton = qt.QPushButton(self)
+        self.previousButton.setIcon(icon_previous)
+        self.lineEdit = qt.QLineEdit(self)
+        self.lineEdit.setFixedWidth(self.lineEdit.fontMetrics().width('%05d' % n))
+        validator = qt.QIntValidator(1, n, self.lineEdit)
+        self.lineEdit.setText("1")
+        self._oldIndex = 0
+        """0-based index"""
+        self.lineEdit.setValidator(validator)
+        self.label = qt.QLabel(self)
+        self.label.setText("of %d" % n)
+        self.nextButton = qt.QPushButton(self)
+        self.nextButton.setIcon(icon_next)
+        self.lastButton = qt.QPushButton(self)
+        self.lastButton.setIcon(icon_last)
+
+        self.mainLayout.addWidget(HorizontalSpacer(self))
+        self.mainLayout.addWidget(self.firstButton)
+        self.mainLayout.addWidget(self.previousButton)
+        self.mainLayout.addWidget(self.lineEdit)
+        self.mainLayout.addWidget(self.label)
+        self.mainLayout.addWidget(self.nextButton)
+        self.mainLayout.addWidget(self.lastButton)
+        self.mainLayout.addWidget(HorizontalSpacer(self))
+
+        self.firstButton.clicked.connect(self._firstClicked)
+        self.previousButton.clicked.connect(self._previousClicked)
+        self.nextButton.clicked.connect(self._nextClicked)
+        self.lastButton.clicked.connect(self._lastClicked)
+        self.lineEdit.editingFinished.connect(self._textChangedSlot)
+
+    def _firstClicked(self):
+        """Select first/lowest frame number"""
+        self.lineEdit.setText("%d" % self.lineEdit.validator().bottom())
+        self._textChangedSlot()
+
+    def _previousClicked(self):
+        """Select previous frame number"""
+        if self._oldIndex >= self.lineEdit.validator().bottom():
+            self.lineEdit.setText("%d" % self._oldIndex)
+            self._textChangedSlot()
+
+    def _nextClicked(self):
+        """Select next frame number"""
+        if self._oldIndex < (self.lineEdit.validator().top() - 1):
+            self.lineEdit.setText("%d" % (self._oldIndex + 2))     # why +2?
+            self._textChangedSlot()
+
+    def _lastClicked(self):
+        """Select last/highest frame number"""
+        self.lineEdit.setText("%d" % self.lineEdit.validator().top())
+        self._textChangedSlot()
+
+    def _textChangedSlot(self):
+        """Select frame number typed in the line edit widget"""
+        txt = self.lineEdit.text()
+        if not len(txt):
+            self.lineEdit.setText("%d" % (self._oldIndex + 1))
+            return
+        new_value = int(txt) - 1
+        if new_value == self._oldIndex:
+            return
+        ddict = {
+            "event": "indexChanged",
+            "old": self._oldIndex + 1,
+            "new": new_value + 1,
+            "id": id(self)
+        }
+        self._oldIndex = new_value
+        self.sigIndexChanged.emit(ddict)
+
+    def setRange(self, first, last):
+        """Set minimum and maximum frame numbers"""
+        return self.setLimits(first, last)
+
+    def setLimits(self, first, last):
+        """Set minimum and maximum frame numbers"""
+        bottom = min(first, last)
+        top = max(first, last)
+        self.lineEdit.validator().setTop(top)
+        self.lineEdit.validator().setBottom(bottom)
+        self._oldIndex = bottom - 1
+        self.lineEdit.setText("%d" % (self._oldIndex + 1))
+        self.label.setText(" limits = %d, %d" % (bottom, top))
+
+    def setNFrames(self, nframes):
+        """Set minimum=1 and maximum frame numbers"""
+        bottom = 1
+        top = nframes
+        self.lineEdit.validator().setTop(top)
+        self.lineEdit.validator().setBottom(bottom)
+        self._oldIndex = bottom - 1
+        self.lineEdit.setText("%d" % (self._oldIndex + 1))
+        self.label.setText(" of %d" % top)
+
+    def getCurrentIndex(self):
+        """Get 1-based index"""
+        return self._oldIndex + 1
+
+    def setValue(self, value):
+        """Set 1-based frame index
+
+        :param int value: Frame number"""
+        self.lineEdit.setText("%d" % value)
+        self._textChangedSlot()
+
+
+class HorizontalSliderWithBrowser(qt.QAbstractSlider):
+    """Frame browser widget, a :class:`FrameBrowser` widget and a slider,
+    to select a frame in a stack of images."""
+    sigIndexChanged = qt.pyqtSignal(object)
+
+    def __init__(self, parent=None):
+        qt.QAbstractSlider.__init__(self, parent)
+        self.setOrientation(qt.Qt.Horizontal)
+
+        self.mainLayout = qt.QHBoxLayout(self)
+        self.mainLayout.setContentsMargins(0, 0, 0, 0)
+        self.mainLayout.setSpacing(2)
+
+        self._slider = qt.QSlider(self)
+        self._slider.setOrientation(qt.Qt.Horizontal)
+
+        self._browser = FrameBrowser(self)
+
+        self.mainLayout.addWidget(self._slider)
+        self.mainLayout.addWidget(self._browser)
+
+        self._slider.valueChanged[int].connect(self._sliderSlot)
+        self._browser.sigIndexChanged.connect(self._browserSlot)
+
+    def setMinimum(self, value):
+        """Set minimum frame number"""
+        self._slider.setMinimum(value)
+        maximum = self._slider.maximum()
+        if value == 1:
+            self._browser.setNFrames(maximum)
+        else:
+            self._browser.setRange(value, maximum)
+
+    def setMaximum(self, value):
+        """Set maximum frame number"""
+        self._slider.setMaximum(value)
+        minimum = self._slider.minimum()
+        if minimum == 1:
+            self._browser.setNFrames(value)
+        else:
+            self._browser.setRange(minimum, value)
+
+    def setRange(self, first, last):
+        """Set minimum/maximum frame numbers"""
+        self._slider.setRange(first, last)
+        self._browser.setRange(first, last)
+
+    def _sliderSlot(self, value):
+        """Emit selected frame number when slider is activated"""
+        self._browser.setValue(value)
+        self.valueChanged.emit(value)
+
+    def _browserSlot(self, ddict):
+        """Emit selected frame number when browser state is changed"""
+        self._slider.setValue(ddict['new'])
+
+    def setValue(self, value):
+        """Set frame number
+
+        :param int value: Frame number"""
+        self._slider.setValue(value)
+        self._browser.setValue(value)
+
+    def value(self):
+        """Get selected frame number"""
+        return self._slider.value()
